# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TestSimply API
  version: 1.0.0
  description: REST API for managing projects, cases, plans, executions, and automated imports.
servers:
  - url: https://testsimplify.local
paths:
...
  /api/testcases:
    get:
      summary: List test cases
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Test case collection
    post:
      summary: Create test case
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseInput'
      responses:
        '201':
          description: Created
...
  /api/testcases/import:
    post:
      summary: Import test cases from CSV
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Imported
  /api/testcases/export:
    get:
      summary: Export test cases as CSV
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/testplans:
    get:
      summary: List test plans
      responses:
        '200':
          description: List of plans
    post:
      summary: Create test plan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestPlanInput'
      responses:
        '201':
          description: Created
...
  /api/import/junit:
    post:
      summary: Import JUnit XML
      parameters:
        - in: query
          name: projectKey
          schema:
            type: string
        - in: query
          name: execKey
          schema:
            type: string
        - in: query
          name: autoCreateCases
          schema:
            type: boolean
        - in: query
          name: createExecutionIfMissing
          schema:
            type: boolean
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                report:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imported
  /api/import/cucumber:
    post:
      summary: Import Cucumber JSON
      parameters:
        - in: query
          name: projectKey
          schema:
            type: string
        - in: query
          name: execKey
          schema:
            type: string
        - in: query
          name: autoCreateCases
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
      responses:
        '200':
          description: Imported
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestCaseInput:
      type: object
      required: [projectId, key, name]
      properties:
        projectId:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        tags:
          type: array
          items:
            type: string
...
    Step:
      type: object
      properties:
        action:
          type: string
        expected:
          type: string
    TestPlanInput:
      type: object
      required: [projectId, key, name]
      properties:
        projectId:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        environments:
          type: array
          items:
            type: string
...